requirements.txt
openai>=1.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tqdm>=4.66.0
